+++ R/reexports.R
@@ -62,6 +62,30 @@ dplyr::summarize
 #' @export
 dplyr::do
 
+#' @importFrom dplyr full_join
+#' @export
+dplyr::full_join
+
+#' @importFrom dplyr inner_join
+#' @export
+dplyr::inner_join
+
+#' @importFrom dplyr left_join
+#' @export
+dplyr::left_join
+
+#' @importFrom dplyr right_join
+#' @export
+dplyr::right_join
+
+#' @importFrom dplyr semi_join
+#' @export
+dplyr::semi_join
+
+#' @importFrom dplyr anti_join
+#' @export
+dplyr::anti_join
+
 #' @importFrom dplyr %>%
 #' @export
 dplyr::`%>%`